% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg.R
\name{aggreg}
\alias{aggreg}
\title{Aggregation function for Ecosystem Functional Properties}
\usage{
aggreg(x, aggregation.time, aggregation.metric, dates,
  overlapping = NULL, probs)
}
\arguments{
\item{x}{EFPs to be aggregated.}

\item{aggregation.time}{Can be "day", "month", "year", growing season "gs". if it's
a number it's considered as a time window in number of days. If "optim" the optimum number of days will be
estimate to reduce the sd of each chunck.}

\item{aggregation.metric}{Can be "mean", "max", "min", "median", and "quant". if "quant" a number between 0 and 1 need to be provided in prob parameter.}

\item{dates}{A vector of class "Date" of the same length of x.}

\item{overlapping}{Can be "NULL" or a number. If it's a number equivale to the parameter by of the \cite{\link[zoo]{rollapply}} function.}

\item{prob}{Only used if aggregation.metric is "quant", a number between 0 and 1. By default 0.9.}
}
\description{
Aggregation function for Ecosystem Functional Properties
}
