% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iAAFTSurrogateEnsembles.R
\name{.iAAFTSurrogateEnsemble}
\alias{.iAAFTSurrogateEnsemble}
\title{Get random ensemble time series}
\usage{
.iAAFTSurrogateEnsemble(x, nsamples = 10)
}
\arguments{
\item{x}{input time series}

\item{nsamples}{number of surrogate time series}
}
\value{

}
\description{
Get random ensemble time series
}
\examples{

}
\author{
Fabian Gans, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de
}
