% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofun.R
\docType{package}
\name{ecofunr}
\alias{ecofunr}
\alias{ecofunr-package}
\title{ecofunr: A package to retrive \strong{Eco}system \strong{Fun}ctional properties from Eddy covariance data.}
\description{
The \code{ecofunr} package provide different functions to estimate several ecosystem functional properties as:
\itemize{
\item Carbon use efficiency (CUE)
\item ....
}
}
\details{
The package also provide different aggregation techniques to compare different ecosystems....
}
