% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalSens.R
\name{evalSens}
\alias{evalSens}
\title{Evaluating the SCAPE performance}
\usage{
evalSens(SCAPE_res, Rb = NA)
}
\arguments{
\item{SCAPE_res}{list: ouput from a successful run of a getQ10, getArrhenius or getLloydTaylor function}

\item{Rb}{vector: in case it applies: an Rb time series}
}
\description{
Evaluating the SCAPE performance
}
\details{
Function to properly evaluating the model based on the SCAPE estimated sensitivities.
A straight forward evaluation of the SCAPE predictions e.g. via the RMSE or other error metrics are
possible, these could be overoptimistic. The reason is that the time varying basal respiration is
extracted as part of the original observations. Hence, a model that includes this part of the signal
is hence comparing a fraction of the signal with itself. The evaluation wrapper uses the output of
the get*** model and performs the evaluation based on the spectrally decomposed signals
(i.e. in frequency ranges where Rb does not play a direct role), using the same spectral method,
parameterization, and surrogate setting. Metrics used here are
}
\seealso{
\code{\link[efps]{getQ10}}, \code{\link[efps]{getLloydTaylor}}, \code{\link[efps]{getArrhenius}}
}
\author{
Fabian Gans, Miguel D. Mahecha, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de mmahecha@bgc-jena.mpg.de
}
