% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_normalized.R
\name{gpp_normalized}
\alias{gpp_normalized}
\title{Normalized GPP}
\usage{
gpp_normalized(GPP, fparc = 0.95, fpar, theta, na.rm = T)
}
\arguments{
\item{GPP}{GPP time series. Object of the class "vector"}

\item{fparc}{fpar normalized at 0.95 (units are missing)}

\item{fpar}{fpar time series. Object of the class "fpar"}

\item{theta}{zenith angle during the solsctice}

\item{na.rm}{}
}
\description{
Normalized GPP
}
