% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gppsat.R
\name{gppsat}
\alias{gppsat}
\title{GPPsat}
\usage{
gppsat(GPP, Radiation, saturation = 1500, method = "NRHLR",
  method_optim = "L-BFGS-B", Amax = "quantile", probs = 0.9,
  alfa = 0.5, Rd = 0, conv = 0.1, modelling_effiency = 0.4,
  ts = T, overlap = T, window_method = "center", length_day = 48,
  width = 5)
}
\arguments{
\item{GPP}{GPP time series. Object of class "vector".}

\item{Radiation}{Radiation (PAR or APAR) time series. Object of class "vector".}

\item{saturation}{Saturation of the radiation. By default 1500. See Details.}

\item{method}{Light response curve model. "NRHLR" or "RHLR". See Details.}

\item{method_optim}{See Details.}

\item{Amax}{Plateau of the light response. Could be "quantile" to use GPP quantile or a number.}

\item{probs}{If Amax is "quantile" probability must be provied (Number between 0-1). By default 0.90.}

\item{alfa}{Inital slope of the light response curve. By default 0.5.}

\item{Rd}{Non-linear response curve intercept. By default 0.}

\item{conv}{Curvature paramenter. Ranging from 0 to 1. By default 0.1}

\item{modelling_effiency}{Quality control for the model evaluation. A number between 0 and 1. By default 0.4}

\item{ts}{Time series output. True or False. By default TRUE. See Details.}

\item{overlap}{Rolling moving window type. Could be TRUE or FALSE. See Details.}

\item{window_method}{Method used to generate the time series. Could be "left", "center", "right". See details.}

\item{length_day}{Number of measures per day. By default 48}

\item{width}{Window width (days). Object of class "number" or "vector". See Details.}
}
\value{
If ts = TRUE the result is a vector. If ts = FALSE the result is a single value.
}
\description{
Ecosystem Photosynthetic Capacity with light saturated
}
\details{
The Ecosystem Photosynthetic Capacity represents the ecosystem potential to uptake CO2 from ecosystem.
}
\examples{



}
