% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lue.R
\name{lue}
\alias{lue}
\title{LUE}
\usage{
lue(GPP, APAR, method = "max", probs = 0.9, ts = T, overlap = T,
  length_day = 48, width = 5, window_method = "center")
}
\arguments{
\item{GPP}{Object of the class 'vector'. Matrix and dataframes are coerced as vectors.}

\item{APAR}{Object of class 'vector'. Matrix and dataframes are coerced as vectors.}

\item{method}{Could be 'max', 'mean' or 'quantile'. By default 'max'}

\item{probs}{if method = 'quantile' a probability need be provided. A number between 0 and 1.}

\item{ts}{output time series. Could be TRUE or FALSE. By default TRUE}

\item{overlap}{if ts = TRUE, overlap must be provided. Could be TRUE or FALSE. By default TRUE}

\item{length_day}{if ts = TRUE. lenght day must be provided. Number of measures per day. By default 48.}

\item{width}{window width (days). Must be an integer number.}

\item{window_method}{if overlap = TRUE, a window method must be provided. Could be 'right', 'center', 'left'. By default 'center' See Details.}
}
\description{
Light use effiency function.
}
