% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotensembles.R
\name{plotensembles}
\alias{plotensembles}
\title{plot the decomposed series}
\usage{
plotensembles(result)
}
\arguments{
\item{result}{list: Results list generated by getQ10}
}
\value{

}
\description{
Function that creates plots of the surrogates of the low-freuquency component of the respiration and temprature data
}
