% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictR.R
\name{predictR}
\alias{predictR}
\title{Predict respiration}
\usage{
predictR(Rb, S, tau, lag = 0)
}
\arguments{
\item{Rb}{numeric vector: basal respiration}

\item{S}{numeric: Sensitivity}

\item{tau}{numeric vector: converted temperature time series}

\item{lag}{}
}
\description{
Calculate respiration for given basal respiration and Q10 value
}
\details{
This function can be useful to calculate predicted respiration for surrogate and time-lagged SCAPE output data
}
\examples{

}
\author{
Fabian Gans, Miguel D. Mahecha, MPI BGC Jena, Germany, fgans@bgc-jena.mpg.de mmahecha@bgc-jena.mpg.de
}
