% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2_efps.R
\name{sentinel2.efps}
\alias{sentinel2.efps}
\title{Predicting Ecosystem Functional Properties using Sentinel-2}
\usage{
sentinel2.efps(sentinel2_product, product = "gpp")
}
\arguments{
\item{sentinel2_product}{The path to a Sentinel-2 L2A product downloaded from https://scihub.copernicus.eu/ or generated with sen2cor. (String)}

\item{product}{Can be "gpp", "gppsat", or "all".}
}
\value{
An object of class raster if product = "gpp" or "gppsat".If class = "all" the output is a stack raster.
}
\description{
This function use pre-trained regression trees to predict GPP and GPPsat.
}
