% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{EFPs temporal series}
\usage{
timeseries(..., data.day = 48, width = 5, overlap = T,
  method = "center")
}
\arguments{
\item{data.day}{A number. number of data per day.}

\item{width}{A number or a vector. Temporal window width. See Details.}

\item{overlap}{Specify if the moving window is overlaped or not could be TRUE or FALSE. See Details}

\item{method}{method used to generated the temporal series could be "center" "right" "left". See Details}

\item{fun}{EFPs function.}
}
\description{
Generate a temporal series.
}
\details{
..., is a dataframe with n columns where each column represent the variables necesaries to
calculate the EFPs.
}
