% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gs.R
\name{ts_gs}
\alias{ts_gs}
\title{Determine the number of growing seasons in a temporal series.}
\usage{
ts_gs(x, max.ft = 5)
}
\arguments{
\item{x}{a vector that represent a temporal serie}

\item{max.ft}{the maximum coeffient that will be used to trace the number of growing seassons}
}
\description{
Determine the number of growing seasons in a temporal series.
}
