% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wue.R
\name{wue}
\alias{wue}
\title{Water Use Efficiency}
\usage{
wue(GPP, VPD, ET, Gs, method = "wue", dates, aggregation.time = "NULL",
  aggregation.metric, overlapping = F, prob = 0.9)
}
\arguments{
\item{GPP}{time series. Object of class "vector".}

\item{VPD}{time series. Object of class "vector".}

\item{ET}{time series. Object of class "vector". See: \code{\link[bigleaf]{LE.to.ET}}.}

\item{Gs}{bulk surface conductance. Object of class "vector"}

\item{method}{WUE index. By default "wue". Options: "wue", "iwue", "Iwue", "uwue", "all". See Details for further explanaition.}

\item{dates}{a vector of class "Date" of the same length of GPP.}

\item{aggregation.time}{Can be "NULL", "day", "month", "year", growing season "gs". if it's
a number it's considered as a time window in number of days. If "optim" the optimum number of days will be estimate to reduce the sd of each chunck. See \code{\link[ecofunr]{aggreg}}.}

\item{aggregation.metric}{Can be "mean", "max", "min", "median", and "quant". if "quant" a number between 0 and 1 need to be provided in prob parameter}

\item{overlapping}{Can be "NULL" or a number. If it's a number equivale to the parameter by of the \cite{\link[zoo]{rollapply}} function.}

\item{prob}{Only used if aggregation.metric is "quant", a number between 0 and 1. By default 0.9.}
}
\value{
An object of type "data.frame" if wue.v = "all". Otherwise an object of the class "vector".
}
\description{
This function estimate different Water use efficiency index.
Water use efficiency
}
\examples{

}
\references{
\insertAllCited{}
}
