% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue.R
\name{cue}
\alias{cue}
\title{Carbon Use Efficiency}
\usage{
cue(GPP, COMP, method = "plants")
}
\arguments{
\item{GPP}{Gross Primary Production (units)}

\item{COMP}{Can be: NEP, NPP, Rb (see XXX Reddyproc function)}

\item{method}{Can be "plants", "ecosystem", "apparent"}
}
\description{
testing
}
\section{Methods}{


\strong{plants}
Carbon Use efficiency (CUE) provides a measure of what fraction of total carbon assimilation becomes
incorported into new tissues. \insertCite{chambers_respiration_2004}

\deqn{CUE = \frac{NPP}{GPP}}{CUE = NPP / GPP}

\strong{ecosystem} \insertCite{fernandez-martinez_nutrient_2014}:

\deqn{CUE = \frac{NEP}{GPP}}{CUE = NEP / GPP}

\strong{apparent} (Check literature ask to Mirco)

\deqn{CUE = 1 - \frac{Rb}{GPP}}{CUE = 1 - (Rb / GPP)}

For a complete review about all CUE metrics at different spatial and temporal scales check \insertCite{manzoni_reviews_2018}
}

\references{
\insertAllCited{}
}
