% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofun.R
\docType{package}
\name{ecofunr}
\alias{ecofunr}
\title{ecofunr: A package to retrive \strong{Eco}system \strong{Fun}ctional properties from Eddy covariance data.}
\description{
The \code{ecofunr} package provide different functions to derive ecosystem functional properties from FLUXNET data as:
\itemize{
\item Carbon use efficiency (cue)
\item Water use efficiency (wue)
\item Evaporative Fraction (ef)
\item Ecosystem Photosynthetic capacity at saturate light (GPPsat)
}
}
\details{
The package also provide different aggregation techniques to compare different ecosystems....
}
