% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef.R
\name{ef}
\alias{ef}
\title{Evaporative Fraction}
\usage{
ef(LE, H, dates, aggregation.time = "NULL", aggregation.metric,
  overlapping = F, prob = 0.9)
}
\arguments{
\item{LE}{Latent heat (units).}

\item{H}{Sensible heat (units).}

\item{dates}{Dates vector.}

\item{prob}{}
}
\description{
XXXXX. \insertCite{nichols_evaluation_1993}{ecofunr}.
}
\details{
Evaporative fraction is the ration between Latent heat and the sum of sensible heat and latent heat.....
}
\examples{


}
\references{
\insertAllCited{}
}
