% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpp_norm.R
\name{gpp_normalized}
\alias{gpp_normalized}
\title{Normalized GPP}
\usage{
gpp_normalized(GPP, fparc = 0.99, fpar, theta, dates, na.rm = T)
}
\arguments{
\item{GPP}{GPP time series (half-hourly). Object of the class "vector"}

\item{fparc}{fpar normalized at (by default) 0.99}

\item{fpar}{fpar time series (daily). Object of the class "vector"}

\item{theta}{zenith angle during the solstice}

\item{dates}{A vector with half-hourly dates of class "Date" or class "numeric" (ISO timestamp). In FLUXNET correspond to the column "TIMESTAMP_START"}

\item{na.rm}{}
}
\value{
A vector of class "numeric".
}
\description{
Normalized GPP
}
\details{
\strong{Normalized Canopy photosythesis} computed as in \insertCite{kergoat_nitrogen_2008-1}{ecofunr}:

\deqn{GPP* = GPPmax*\frac{fPAR* cos(\theta = 0)}{fParc * cos(\theta_{s})}}{GPP* = GPPmax * (fPARc * cos(\theta = 0)) / (fPAR * cos(\theta_{s}))}

Where:

GPPmax: Quantile 0.985-0.99 of the half-hourly GPP for the period of interest (per day).
fPAR: Fraction of absorbed PAR.
fPARc: Reference fPAR ofr closed canopy conditions (by default 0.95).
\theta_{s}: The sun elevation at solstice.
}
\examples{

#' @references
\\insertAllCited{}


}
