% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wue.R
\name{wue}
\alias{wue}
\title{Water Use Efficiency}
\usage{
wue(
  GPP,
  VPD,
  ET,
  Gs,
  method = "wue",
  dates,
  aggregation.time = NULL,
  aggregation.metric = "median",
  overlapping = NULL,
  probs = 0.9
)
}
\arguments{
\item{GPP}{time series. Object of class "vector" (umol CO2 m-2 s-1).}

\item{VPD}{time series. Object of class "vector" (kPa).}

\item{ET}{time series. Object of class "vector". See: \code{\link[bigleaf]{LE.to.ET}} (kg m-2 s-1).}

\item{Gs}{bulk surface conductance. Object of class "vector"}

\item{method}{WUE index. By default "wue". Options: "wue", "iwue", "Iwue", "uwue", "all". See Details for further explanaition.}

\item{dates}{a vector of class "Date" with the same length of GPP.}

\item{aggregation.time}{Can be "NULL", "day", "month", "year", growing season "gs". if it's
a number it's considered as a time window in number of days. If "optim" the optimum number of days will be estimate to reduce the sd of each chunck. See \code{\link[ecofunr]{aggreg}}.}

\item{aggregation.metric}{Can be "mean", "max", "min", "median", and "quantile". if "quantile" a number between 0 and 1 need to be provided for prob parameter.}

\item{overlapping}{Can be "NULL" or a number. If it's a number equivale to the parameter by of the \cite{\link[zoo]{rollapply}} function.}

\item{prob}{Only used if aggregation.metric is "quant", a number between 0 and 1. By default 0.9.}
}
\value{
An object of type "data.frame" if method = "all". Otherwise an object of the class "vector".
}
\description{
This function estimate different metrics of Water Use Efficiency.
}
\details{
\strong{Water Use Efficiency (wue) \insertCite{law_environmental_2002}{ecofunr}}

\deqn{WUE = \frac{GPP (gCm^{-2}s^{-1})}{ET (kg H_{2}Om^{-2}s^{-1})}}{WUE = GPP / ET}

\strong{Intrinsic Water use efficiency (iwue) See: \insertCite{beer_temporal_2009-1}{ecofunr}}

\deqn{iWUE = \frac{GPP}{G_{s}}}{iWUE = GPP / Gs}

where Gs is the bulk surface conductance. See \code{\link[bigleaf]{surface.conductance}}

\strong{Inherent Water Use Efficiency (Iwue) \insertCite{beer_temporal_2009-1}{ecofunr}}

\deqn{IWUE = \frac{GPP*VPD}{ET}}{IWUE = GPP * VPD / ET}

\strong{Underlyng Water Use Efficiency (uwue)} \insertCite{zhou_effect_2014}{ecofunr}

\deqn{uWUE = \frac{GPP * \sqrt{VPD}}{ET}}{uWUE = GPP * sqrt(VPD) / ET}

\strong{Undelying Water Use Efficiency using transpiration from TEA algorithm (teauWUE)} \insertCite{nelson_coupling_2018-1}{ecofunr}

\deqn{teauWUE = \frac{GPP * \sqrt{VPD}}{T}}{teauWUE = GPP * sqrt(VPD) / T}

where T is Transpiration derived from TEA algorithm \insertCite{nelson_coupling_2018-1}{ecofunr}
}
\examples{

}
\references{
\insertAllCited{}
}
